Things are working now.
Blocks are being mined, transactions can be sent, and this is a good
start. I needed to include the signers/miners in the "extradata" field
in the genesis file. But now I want to see how I can add/propose more
signers and add them to the network. Also need to figure out how to
deploy smart contracts on the network.