Messing around with config variables.
Added two scripts to automate process of adding/removing signers from
network. Could be improved, but works well for now. I'm surprised that
it doesn't seem like the epoch is being used at all as a delineation for
when to apply the result of the votes to the network. It seems like once
I run the script to add a signer to the network, it happens almost
immediately. Same with removing a signer from the network.