Updated README and script to allow for switching between IBFT and Clique
IBFT isn't fully working yet though. There is an issue where miners are
not "authorized" so no blocks get mined. I have some ideas I'm still
working on to figure out how to fix this. But I wanted to keep the
script the same as much as possible. It should be possible to pretty
much do exactly the same things for the majority of the setup process
and just change the underlying consensus being used.