Updated main script to allow the proper bootnode bin to be specified
Updated the startIBFTNetwork and startCliqueNetwork scripts to account
for this new parameter.

Also updated the README with some instructions for terminating geth and
bootnode processes that sometimes continue running after the main script
finishes running.

Last thing I added was a function decode the RLP-encoded IBFT extraData
field to ensure that it's being properly encoded (which it is). It
didn't seem like that was the issue, but I figured it's good to have in
place anyways.

The issue with the IBFT network right now is that nodes are hitting an
error saying "Failed to decode message from payload." The address it's
showing isn't the address of any node on the local network, and it seems
like it's getting it from the node's ID which can be found by attaching
to a node ("geth attach path/to/geth.ipc") and running admin.nodeInfo
and looking at the ID. Not sure how this is generated, but IBFT thinks
that the node's address is in this ID and it doesn't actually match with
the node's proper address.